RFC-061 Phase 3: delete start_compiled + Slots extractor + bridge contract test
Reduced-scope Phase 3 (vs. plan): ships the architectural commitment
without the slot-store retirement. Plan envisioned deleting the entire
adopted-DOM bridge in one PR; Phase 3a delivers the surgical subset
that tests actually allow.

Deletions:
  - `walker::start_compiled` (~57 LOC, the app-level discovery entry).
    Body lives on as the private `mount_pp_app_subtree(host)` helper
    in `app.rs` — same shape (querySelectorAll union, mount each via
    `mount_child_component`, finalize via `finalize_compiled_subtree`,
    register pp-outlet) but now scoped to the `[pp-app]` subtree.
  - `Slots` extractor (`lifecycle.rs:421-427` + From<LifecycleContext>
    impl + umbrella re-exports). Zero in-tree callers per `rg : Slots`;
    downstream consumers needing slot-name introspection rebuild on
    compile-time `#[slot]` metadata in their own crate.
  - `crates/pocopine/tests/adopted_dom_contract.rs` (the 4 contract
    cases pinning the bridge). RFC 061 §6 explicitly calls for this
    deletion — the contract is gone with `start_compiled`.

Un-deprecations (these survive Phase 3 as the typed compiled-mount
+ pp-for body fallback + pp-init paths; the Phase 1 `#[deprecated]`
came off):
  - `walker::mount_child_component`
  - `walker::mount_child_component_with_slots`
  - `walker::mount_adopted_components`
  - `walker::defer_init_on`

Test fixture migration: each test crate's `mount(html)` helper +
3 inline `start_compiled` callers in `template_plan.rs` switched to
a shared `mount_registered_tags(host)` helper that mirrors the old
start_compiled body. The 106 + 30 individual test bodies didn't
need changes — the helpers absorbed the swap.

App::run path: `[pp-app]` discovery (Phase 2) + new
`mount_pp_app_subtree` helper (Phase 3) replace the old
`start_compiled(body)` call.

DEFERRED to a follow-on phase:
  - `slots` module + `capture_slots` + `materialize_adopted_slot`
    + the slots::lookup branch in materialize_slot. Pine compounds
    that author raw-HTML slot content (`<pine-tooltip-root>
    <button>...</button></pine-tooltip-root>`) depend on the runtime
    slot store; deletion broke 50 pine tests in a smoke run. Full
    retirement needs ~106 per-test fixture-component migrations
    (each test would author its slot content inside a
    `template_inline` wrapper so the macro emits a compile-time
    slot fragment). That's its own RFC follow-on, not appropriate
    for this PR's scope.
  - `walker.rs → mount.rs` rename. Premature while ~1500 LOC of
    bridge surface (capture_slots, materialize_adopted_slot,
    mount_adopted_components, install_adopted_controllers) still
    lives there. Pairs with the slot-store retirement above.

Acceptance gates passing:
  - pine 106/106
  - pocopine 44 wasm tests across 5 files (adopted_dom_contract gone)
  - cargo clippy --workspace --target wasm32-unknown-unknown --all-targets -- -D warnings clean
  - cargo fmt --all --check clean
  - No stale `#![allow(deprecated)]` annotations in
    pocopine-core consumer files (for_.rs / router.rs /
    templates_plan.rs) — the deprecations they suppressed are
    gone. `walker.rs` keeps its file-level allow because the
    slots module it consumes is still deprecated (deferred).
