+++ crates/pocopine/src/lib.rs
@@ -24,8 +24,8 @@ pub use pocopine_core::{
     ContextMarker, Doc, DomEventName, EffectId, EffectOptions, El, Elapsed, Emit, Handle, HostEl,
     Inject, IsTeleported, LifecycleContext, ListenerHandle, MountEpoch, NearestParent, Parent,
     ParentId, Refs, RegisteredComponent, RegistryError, RegistryErrorKind, RwSignal, Scope,
-    ScopeId, ScopePath, ServerError, ServerResult, Setter, Signal, SignalId, Slots, Store,
-    StoreHandle, SubtreeHandle, TagName, TaskHandle, TeleportHost, TypedEl, Win,
+    ScopeId, ScopePath, ServerError, ServerResult, Setter, Signal, SignalId, Store, StoreHandle,
+    SubtreeHandle, TagName, TaskHandle, TeleportHost, TypedEl, Win,
 };
 #[doc(inline)]
 pub use pocopine_core::{create_context, inject_key};
@@ -82,17 +82,6 @@ pub mod __private {
     // call through the umbrella's stable `::pocopine::__private`
     // namespace. Same shape as the existing `register_template` /
     // `register_row_plans` re-exports above.
-    //
-    // RFC 061 Phase 1 — `defer_init_on`,
-    // `mount_child_component`, and
-    // `mount_child_component_with_slots` are deprecated; the
-    // re-exports stay for the macro's compiled emission path.
-    // Phase 3 deletes them along with the underlying walker
-    // functions. Suppressing the deprecation warning on the
-    // re-export line keeps macro-generated code quiet without
-    // hiding the warning from external callers (they hit the
-    // deprecation on the function itself).
-    #[allow(deprecated)]
     pub use pocopine_core::walker::{
         defer_init_on, finalize_compiled_subtree, fire_mount_post_order, fire_ready_next_tick,
         mount_child_component, mount_child_component_with_slots, release_compiled_subtree,
